Changed self-signed certificates Disabled ssl
