+++ app.py
@@ -55,8 +55,8 @@ db_session.global_init('db/database.sqlite')
 
 
 def main():
-    port = int(os.environ.get('PORT', 443))
-    socket_.run(app, host='0.0.0.0', port=port, keyfile='private.key', certfile='certificate.crt')
+    port = int(os.environ.get('PORT', 80))
+    socket_.run(app, host='0.0.0.0', port=port)
 
 
 def add_admin_panel():
